2.填写标书/点对点应答/requirements.txt
flask>=2.0.0
python-docx>=0.8.11
requests>=2.25.0
werkzeug>=2.0.0